refactor: PageDto 클래스 이동
